+++ src/components/CreatePage.jsx
@@ -1,10 +1,17 @@
 import React from 'react'
 import Form from './Form'
+import PetPage from './PetPage'
+
 
 function CreatePage () {
     return (
         <div>
-            <Form />
+            {/* <h1>From create page</h1> */}
+            {/* <Form /> */}
+            <PetPage/>
+          
+            
+
         </div>
         
     )
