+++ src/components/StatusBar.jsx
