+++ src/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './components/App.jsx';
-import CreatePage from '../components/CreatePage.jsx';
+import CreatePage from './components/CreatePage.jsx';
 // import './scss/_app.scss';
+import { BrowserRouter } from 'react-router-dom';
 
 const root = createRoot(document.getElementById('root'));
 root.render(<App />);
